third_party/WebKit/LayoutTests/fast/selectors/style-sharing-shadow-expected.txt

CONSOLE WARNING: Calling Element.createShadowRoot() for an element which already hosts a shadow root is deprecated. See https://www.chromestatus.com/features/4668884095336448 for more details.
PASS successfullyParsed is true

TEST COMPLETE
PASS true is true
PASS true is true
PASS false is false
PASS false is false
PASS false is false
PASS true is true
PASS false is false
PASS false is false
item 1
item 2
A

B
